# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.AssignmentResultResponse:
    properties:
      exam_id:
        type: string
      exam_name:
        type: string
      score:
        type: number
      submitted_at:
        type: string
    type: object
...
  entity.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  entity.ForgotPasswordResponse:
    properties:
      message:
        type: string
    type: object
  entity.GetAssignmentListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.AssignmentListItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  entity.GetLeaderboardResponse:
    properties:
      exam_id:
        type: string
      exam_name:
        type: string
      items:
        items:
          $ref: '#/definitions/entity.LeaderboardItem'
        type: array
    type: object
...
  entity.LoginRequest:
    properties:
      identifier:
        description: email atau username
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  entity.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/entity.UserResponse'
    type: object
  entity.LogoutResponse:
    properties:
      message:
        type: string
    type: object
...
  entity.RegisterResponse:
    properties:
      message:
        type: string
    type: object
...
  helper.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend service (hexagon/clean architecture).
  title: Widyaprada Backend API
  version: "1.0"
paths:
  /api/v1/assignments:
    get:
      consumes:
      - application/json
      description: Ujian yang user ikuti, filter status, sort batas_waktu
      parameters:
      - description: belum_dikerjakan | sudah_dikerjakan
        in: query
        name: status
        type: string
      - default: jadwal_selesai
        description: jadwal_selesai
        in: query
        name: sort_by
        type: string
      - default: asc
        description: asc | desc
        in: query
        name: sort_order
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetAssignmentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tugas Saya
      tags:
      - assignments
  /api/v1/assignments/{examId}/leaderboard:
    get:
      consumes:
      - application/json
      description: Hanya jika tampilkan_leaderboard=Ya & user peserta. 403 jika privat.
      parameters:
      - description: Exam ID
        in: path
        name: examId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetLeaderboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Leaderboard ujian
      tags:
      - assignments
  /api/v1/assignments/{examId}/result:
    get:
      consumes:
      - application/json
      description: Hasil ujian untuk exam tersebut (hanya peserta)
      parameters:
      - description: Exam ID
        in: path
        name: examId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AssignmentResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hasil ujian user
      tags:
      - assignments
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Minta link reset password ke email
      parameters:
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ForgotPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Forgot Password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan identifier (email/username) dan password
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: 'Keluar dari aplikasi, invalidasi token. Requires Authorization:
        Bearer &lt;token&gt;'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LogoutResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Registrasi mandiri calon peserta (nama, email, nip opsional). Password
        dikirim ke email.
      parameters:
      - description: Register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Register
      tags:
      - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Atur ulang kata sandi dengan token dari email
      parameters:
      - description: Token, password, confirm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Reset Password
      tags:
      - auth
  /api/v1/cbt/attempts/{attemptId}/answers:
    post:
      consumes:
      - application/json
      description: Simpan jawaban (option_id untuk PG/B-S, answer_text untuk Essay)
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      - description: Jawaban
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CBTSaveAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTSaveAnswerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Simpan jawaban
      tags:
      - cbt
  /api/v1/cbt/attempts/{attemptId}/questions:
    get:
      consumes:
      - application/json
      description: Daftar soal untuk attempt
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTListQuestionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar soal
      tags:
      - cbt
  /api/v1/cbt/attempts/{attemptId}/questions/{num}:
    get:
      consumes:
      - application/json
      description: Get soal by nomor (1-based)
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      - description: Nomor soal
        in: path
        name: num
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTQuestionItem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soal per nomor
      tags:
      - cbt
  /api/v1/cbt/attempts/{attemptId}/submit:
    post:
      consumes:
      - application/json
      description: Submit ujian dan hitung skor (PG/B-S otomatis)
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTSubmitResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit ujian
      tags:
      - cbt
  /api/v1/cbt/exams:
    get:
      consumes:
      - application/json
      description: Ujian yang bisa dikerjakan (Diterbitkan, jadwal aktif, user peserta)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTListExamsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar ujian tersedia
      tags:
      - cbt
  /api/v1/cbt/exams/{id}/start:
    post:
      consumes:
      - application/json
      description: Create attempt dan return attempt_id
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mulai ujian
      tags:
      - cbt
  /api/v1/cbt/history:
    get:
      consumes:
      - application/json
      description: Riwayat ujian + nilai user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Riwayat ujian
      tags:
      - cbt
...
swagger: "2.0"